fix: handle multiple models simultaneously
This uses a newer version of the SD client library, which uses
the override_settings parameter on generations instead of
the options endpoint. This allows it to switch models at the
right time, as opposed to "whenever the queue lock is released".

Depends on https://github.com/AUTOMATIC1111/stable-diffusion-webui/pull/6012